Rename api.js to crypto.js
